+++ CHANGELOG.md
@@ -1,3 +1,21 @@
-## 0.0.1
+# Changelog
 
-* Widget HorizontalListView created.
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub/versioning).
+
+## [Unreleased]
+
+## [0.0.2] - 2022-02-02
+
+### Added
+
+- example project;
+- insert size height and width for Widget
+
+## [0.0.1] - 2022-02-01
+
+### Added
+
+- Create the lib horizontal_list with widget HorizontalListView